Made all the floats doubles.

+++ org/bitbuckets/frc2015/RandomConstants.java
@@ -1,7 +1,7 @@
 package org.bitbuckets.frc2015;
 
 public class RandomConstants {
-	public static final float SPEED_K_EXSPMA = .01f;
+	public static final double SPEED_K_EXSPMA = .01f;
 	
-	public static final float ENC_TICK_TO_IN = 8f;
+	public static final double ENC_TICK_TO_IN = 8f;
 }
