+++ org/bitbuckets/frc2015/control/ExampleMotionController.java
@@ -5,26 +5,26 @@ import org.bitbuckets.frc2015.lib.MotionController;
 
 public class ExampleMotionController implements MotionController{
 	/** The sensor value the last time <code>update()</code> was called. **/
-	private float prevSense;
+	private double prevSense;
 	/** The current speed of the <code>SpeedManager</code>. **/
-	private float speed;
+	private double speed;
 
 	@Override
-	public void init(float... inputs) {
+	public void init(double... inputs) {
 		prevSense = inputs[0];
 		speed = 0;
 	}
 	
 	@Override
-	public float getSpeed() {
+	public double getSpeed() {
 		return speed;
 	}
 
 	@Override
-	public float update(float set, float... input) {
-		float sense = input[0];
-		float moved = sense-prevSense; 
-		float error = set - moved;
+	public double update(double set, double... input) {
+		double sense = input[0];
+		double moved = sense-prevSense; 
+		double error = set - moved;
 		
 		speed = set + RandomConstants.SPEED_K_EXSPMA * error;
